Refactored Gizmo
