+++ src/components/canvas/Gizmo/index.ts
@@ -1,0 +1,4 @@
+import { Gizmo } from './Gizmo'
+import { findInteractableParent, getGizmoInfo } from './utils'
+
+export { Gizmo, findInteractableParent, getGizmoInfo }
